refactor(ARCO-304): move woc_client to pkg
