+++ cmd/broadcaster-cli/app/keyset/topup/topup.go
@@ -8,7 +8,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/bitcoin-sv/arc/cmd/broadcaster-cli/helper"
-	"github.com/bitcoin-sv/arc/internal/woc_client"
+	"github.com/bitcoin-sv/arc/pkg/woc_client"
 )
 
 var Cmd = &cobra.Command{
