+++ cmd/broadcaster-cli/app/keyset/utxos/utxos.go
@@ -12,8 +12,8 @@ import (
 	"github.com/spf13/viper"
 
 	"github.com/bitcoin-sv/arc/cmd/broadcaster-cli/helper"
-	"github.com/bitcoin-sv/arc/internal/woc_client"
 	"github.com/bitcoin-sv/arc/pkg/keyset"
+	"github.com/bitcoin-sv/arc/pkg/woc_client"
 )
 
 var Cmd = &cobra.Command{
