+++ cmd/broadcaster-cli/app/utxos/consolidate/consolidate.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/bitcoin-sv/arc/cmd/broadcaster-cli/helper"
 	"github.com/bitcoin-sv/arc/internal/broadcaster"
-	"github.com/bitcoin-sv/arc/internal/woc_client"
+	"github.com/bitcoin-sv/arc/pkg/woc_client"
 )
 
 var Cmd = &cobra.Command{
