+++ cmd/broadcaster-cli/app/utxos/create/create.go
@@ -10,7 +10,8 @@ import (
 
 	"github.com/bitcoin-sv/arc/cmd/broadcaster-cli/helper"
 	"github.com/bitcoin-sv/arc/internal/broadcaster"
-	"github.com/bitcoin-sv/arc/internal/woc_client"
+	"github.com/bitcoin-sv/arc/pkg/woc_client"
+
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
